moving scrape functions
