+++ prusalink/einsy.go
@@ -2,66 +2,8 @@ package prusalink
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/pstrobl96/prusa_exporter/config"
 	"github.com/rs/zerolog/log"
 )
-
-func getEinsyResponse(printer config.Printers) (Version, Files, Job, Printer, Info, Settings, Cameras, error) {
-	var (
-		version     Version
-		files       Files
-		job         Job
-		printerData Printer
-		info        Info
-		settings    Settings
-		cameras     Cameras
-		err         error
-	)
-
-	version, err = GetVersion(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting version" + err.Error())
-	}
-
-	files, err = GetFiles(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting files" + err.Error())
-	}
-
-	job, err = GetJob(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting job" + err.Error())
-	}
-
-	printerData, err = GetPrinter(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting printer" + err.Error())
-	}
-
-	info, err = GetInfo(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting info" + err.Error())
-	}
-
-	settings, err = GetSettings(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting settings" + err.Error())
-	}
-
-	cameras, err = GetCameras(printer)
-
-	if err != nil {
-		log.Error().Msg("Error getting cameras" + err.Error())
-	}
-
-	return version, files, job, printerData, info, settings, cameras, err
-}
 
 type einsyCollector struct {
 	printerNozzleTemp         *prometheus.Desc
